Create <T> Generic in simple words.md
A conventional letter refers to Type and points out the Generic concept in Java. It is in play whenever we want to pass a parameter for a class, method, or interface later. Keep in mind, It is conditioned to be in charge of Objects not primitives because Generics are going to be used as a feature at compile time.
Reference : 
1. https://techwithmaddy.com/what-does-lesstgreater-mean-in-java